requirements.txt
fire==0.4.0
functorch==0.1.1
gin-config==0.5.0
matplotlib==3.5.1
numpy ==1.19.2
pandas==1.4.2
scikit-learn==1.0.2
scipy==1.7.3
tensorboard==2.8.0
torch==1.11.0
TorchOpt==0.4.1
tqdm==4.62.3
einops==0.4.1